Moved dashboard reducer and actions to separate files
Just a little refactor to move the dashboard reducer and the dashboard
actions (along with associated tests) to separate files. Much fixing of
stubs was required.

I'm doing this as the first part of #2672
pr #2680
