+++ static/js/reducers/course_enrollments_test.js
@@ -15,6 +15,7 @@ import {
 } from '../actions/course_enrollments';
 import * as api from '../lib/api';
 import * as actions from '../actions';
+import * as dashboardActions from '../actions/dashboard';
 import rootReducer from '../reducers';
 
 describe('enrollments', () => {
@@ -37,7 +38,7 @@ describe('enrollments', () => {
 
       fetchCoursePricesStub = sandbox.stub(actions, 'fetchCoursePrices');
       fetchCoursePricesStub.returns({type: "fake"});
-      fetchDashboardStub = sandbox.stub(actions, 'fetchDashboard');
+      fetchDashboardStub = sandbox.stub(dashboardActions, 'fetchDashboard');
       fetchDashboardStub.returns({type: "fake"});
     });
